requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asttokens=2.0.5=pypi_0
async-generator=1.10=pypi_0
attrs=21.4.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.10.0=pyh06a4308_0
blas=1.0=mkl
bottleneck=1.3.4=py39h080aedc_0
brotlipy=0.7.0=py39h2bbff1b_1003
bs4=4.10.0=hd3eb1b0_0
ca-certificates=2022.3.29=haa95532_0
cachetools=5.0.0=pypi_0
certifi=2021.10.8=py39haa95532_2
cffi=1.15.0=py39h2bbff1b_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.4=pypi_0
cryptography=36.0.0=py39h21b164f_0
db-dtypes=1.0.0=pypi_0
decorator=5.1.1=pypi_0
executing=0.8.3=pypi_0
gcloud=0.18.3=pypi_0
google-api-core=2.7.3=pypi_0
google-api-python-client=2.46.0=pypi_0
google-auth=2.6.6=pypi_0
google-auth-httplib2=0.1.0=pypi_0
google-cloud=0.34.0=pypi_0
google-cloud-bigquery=3.0.1=pypi_0
google-cloud-bigquery-storage=2.13.1=pypi_0
google-cloud-core=2.3.0=pypi_0
google-cloud-storage=2.3.0=pypi_0
google-crc32c=1.3.0=pypi_0
google-resumable-media=2.3.2=pypi_0
googleapis-common-protos=1.56.0=pypi_0
grpcio=1.44.0=pypi_0
grpcio-status=1.44.0=pypi_0
h11=0.13.0=pypi_0
httplib2=0.20.4=pypi_0
idna=3.3=pyhd3eb1b0_0
intel-openmp=2021.4.0=haa95532_3556
ipython=8.2.0=pypi_0
jedi=0.18.1=pypi_0
lxml=4.8.0=pypi_0
matplotlib-inline=0.1.3=pypi_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.1=py39h277e83a_0
mkl_random=1.2.2=py39hf11a4ad_0
numexpr=2.8.1=py39hb80d3ca_0
numpy=1.21.5=py39h7a0a035_1
numpy-base=1.21.5=py39hca35cd5_1
oauth2client=4.1.3=pypi_0
openssl=1.1.1n=h2bbff1b_0
orderedmultidict=1.0.1=pypi_0
outcome=1.1.0=pypi_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.1=py39hd77b12b_1
parso=0.8.3=pypi_0
pickleshare=0.7.5=pypi_0
pip=21.2.4=py39haa95532_0
prompt-toolkit=3.0.29=pypi_0
proto-plus=1.20.3=pypi_0
protobuf=3.20.1=pypi_0
pure-eval=0.2.2=pypi_0
pyaml-env=1.1.5=pypi_0
pyarrow=7.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pysocks=1.7.1=py39haa95532_0
python=3.9.12=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2021.3=pyhd3eb1b0_0
pyyaml=5.4.1=pypi_0
requests=2.27.1=pyhd3eb1b0_0
rsa=4.8=pypi_0
selenium=4.1.3=pypi_0
setuptools=61.2.0=py39haa95532_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.2.0=pypi_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.3.1=pyhd3eb1b0_0
sqlite=3.38.2=h2bbff1b_0
stack-data=0.2.0=pypi_0
traitlets=5.1.1=pypi_0
trio=0.20.0=pypi_0
trio-websocket=0.9.2=pypi_0
tzdata=2022a=hda174b7_0
uritemplate=4.1.1=pypi_0
urllib3=1.26.8=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pypi_0
webdriver-manager=3.5.4=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py39haa95532_0
wincertstore=0.2=py39haa95532_2
wsproto=1.1.0=pypi_0
yaml=0.2.5=he774522_0
